:truck: [#2418] -- move DjangoTemplateValidator to openforms.template package
Now that we have a dedicated package for templating, this is the more
logical place for this validator to live.
