+++ docs/developers/backend/core/templating.rst
@@ -22,6 +22,13 @@ Module functionality
 
 .. autoattribute:: openforms.template.sandbox_backend
 
+Validators
+----------
+
+.. automodule:: openforms.template.validators
+    :members:
+    :exclude-members: deconstruct
+
 Backends
 --------
