+++ src/openforms/forms/models/form.py
@@ -23,8 +23,8 @@ from openforms.payments.registry import register as payment_register
 from openforms.plugins.constants import UNIQUE_ID_MAX_LENGTH
 from openforms.registrations.fields import RegistrationBackendChoiceField
 from openforms.registrations.registry import register as registration_register
+from openforms.template.validators import DjangoTemplateValidator
 from openforms.utils.files import DeleteFileFieldFilesMixin, DeleteFilesQuerySetMixin
-from openforms.utils.validators import DjangoTemplateValidator
 from openforms.variables.constants import FormVariableSources
 
 from ..constants import ConfirmationEmailOptions, SubmissionAllowedChoices
